requirements-txt.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.1.2
Flask-Limiter==2.0.1
